docs(migration-timeline): moved service into shared
