+++ docs_app/src/app/custom-elements/migration-timeline/components/migration-timeline/breaking-change-description-table.component.ts
@@ -1,7 +1,8 @@
 import {Component, Input, Output} from '@angular/core';
 import {Subject} from 'rxjs';
+import {State} from '../../../../shared/state.service';
 import {ClientBreakingChange} from '../../data-access';
-import {State} from '../../utils';
+
 
 interface VMBreakingChangeDescriptionTable {
   breakingChange: ClientBreakingChange;
