+++ docs_app/src/app/custom-elements/migration-timeline/components/migration-timeline/migration-timeline.component.ts
@@ -3,8 +3,7 @@ import {Subject} from 'rxjs';
 import {map} from 'rxjs/operators';
 import {environment} from '../../../../../environments/environment';
 import {ClientMigrationTimelineReleaseItem, parseMigrationReleaseUIDFromString} from '../../data-access';
-
-import {State} from '../../utils';
+import {State} from '../../../../shared/state.service';
 
 export interface MigrationTimelineComponentViewBaseModel {
   releaseList: ClientMigrationTimelineReleaseItem[];
