+++ docs_app/src/app/custom-elements/migration-timeline/components/release-navigation.component.ts
@@ -1,7 +1,7 @@
 import {Component, Input, Output} from '@angular/core';
 import {Subject} from 'rxjs';
 import {ClientMigrationTimelineReleaseItem} from '../data-access';
-import {State} from '../utils';
+import {State} from '../../../shared/state.service';
 
 export interface MigrationTimelineNavigationItem {
   date: Date;
