+++ docs_app/src/app/custom-elements/migration-timeline/data-access/migration-timeline.service.ts
@@ -1,8 +1,7 @@
 import {Injectable} from '@angular/core';
 import {Observable, of, pipe} from 'rxjs';
 import {map, startWith} from 'rxjs/operators';
-
-import {GlobalState} from '../utils';
+import {GlobalState} from '../../../shared/state.service';
 import {ClientMigrationTimelineReleaseItem, compareByReleaseDateAsc, parseClientMigrationTimelineReleaseItem} from './migration-item';
 import {deprecationAndBreakingChangeTimeline} from './migration-timeline-structure';
