+++ docs_app/src/app/shared/operators/index.ts
