+++ docs_app/src/app/shared/operators/stateful.ts
